USB2000spec driver is now working, it works with trigger, and set the integration time correctly

+++ src/sensors/usb2000spec/USB2000Lib.h
@@ -12,7 +12,7 @@
 typedef struct {
     int dataSize;
     int numScans;
-    int intTime;
+    long intTime;
     int baselineMSW;
     int baselineLSW;
     int pixelMode;
